Change the default terminal to [Foot terminal](https://codeberg.org/dnkl/foot), a Wayland-first terminal which is the default for the Sway project. Compared to xfce4-terminal, Foot is faster and more lightweight (both in size and resource usage). It also has native Sixel support, which allows rendering images in the terminal.
I have added a foot.ini which preserves the color scheme that had been set up on xfce4-terminal (with the purple background and transparency), except using the Material color scheme for terminal fonts. The bright colors of the fonts look good on the purple, and are easier to read than the old font colors.

Other changes in this PR:
* Add foot to package lists; remove xfce4-terminal
* Autostart foot-server in Sway config
* Change default terminal in Sway config
* Change Waybar on-click actions to use footclient instead of xfce4-terminal
* Update cheatsheet, keyhint files
* Update Wofi, Thunar configs to use footclient instead of xfce4-terminal